+++ src/components/Router/helpers.ts
@@ -1,0 +1,11 @@
+enum PostNavigationOptions {
+	ScrollTop,
+	ScrollTopSmooth,
+}
+
+enum PostNavigationFocusOptions {
+	FocusBody,
+	FocusHeading,
+}
+
+export { PostNavigationOptions, PostNavigationFocusOptions };